선 접속자도 클라이언트에서 뜨게 하기

+++ Client/Assets/03_Scripts/Network/Player/Prefeb/EntityPrefabAuthoring.cs
@@ -3,7 +3,7 @@ using Unity.Entities;
 
 public struct PlayerInfo : IComponentData
 {
-    public int PlayerID;
+    public ulong PlayerID;
 }
 
 public class EntityPrefabAuthoring : MonoBehaviour
